+++ lib/flame_fuse.dart
@@ -12,4 +12,3 @@ export 'src/hover.dart';
 export 'src/keys.dart';
 export 'src/pointers.dart';
 export 'src/taps.dart';
-export 'src/taps2.dart';